Reviewed by Adele.
        Fixes <rdar://problem/4498338> JavaScriptCore fails to compile for ppc64
        Other 64 bit build fixes.

        * kjs/collector.cpp:
        (KJS::Collector::markOtherThreadConservatively): test for __DARWIN_UNIX03 and use __r1
        * kjs/dtoa.cpp:
        (Bigint::): cast PRIVATE_mem to unsigned to prevent warning
        * bindings/jni/jni_utility.cpp:
        (KJS::Bindings::getJavaVM): cast jniError to long to prevent format warning
        (KJS::Bindings::getJNIEnv): cast jniError to long to prevent format warning
        * bindings/runtime_root.cpp:
        (KJS::Bindings::addNativeReference): cast CFDictionaryGetValue to unsigned long to prevent warning
        (KJS::Bindings::removeNativeReference): cast CFDictionaryGetValue to unsigned long to prevent warning



Canonical link: https://commits.webkit.org/11593@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@13658 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ JavaScriptCore/bindings/jni/jni_utility.cpp
@@ -55,7 +55,7 @@ JavaVM *getJavaVM()
         jvm = jvmArray[0];
     }
     else 
-        fprintf(stderr, "%s: JNI_GetCreatedJavaVMs failed, returned %ld\n", __PRETTY_FUNCTION__, jniError);
+        fprintf(stderr, "%s: JNI_GetCreatedJavaVMs failed, returned %ld\n", __PRETTY_FUNCTION__, (long)jniError);
         
     return jvm;
 }
@@ -69,7 +69,7 @@ JNIEnv *getJNIEnv()
     if ( jniError == JNI_OK )
         return env;
     else
-        fprintf(stderr, "%s: AttachCurrentThread failed, returned %ld\n", __PRETTY_FUNCTION__, jniError);
+        fprintf(stderr, "%s: AttachCurrentThread failed, returned %ld\n", __PRETTY_FUNCTION__, (long)jniError);
     return NULL;
 }
